requirements.txt
fastapi
uvicorn
pydantic
apscheduler
aiosqlite
python-dotenv
requests
schedule
